test(notifier): Update per code review
